requirements.txt
Django==5.1.2
Pillow==10.2.0
python-dotenv==1.0.1
django-debug-toolbar==4.1.0 